requirements.txt
#
#  System requirements for AC-PKI
#

pyOpenSSL >= 19.0.0

# acitoolkit requirements
requests >= 2.16.0
websocket-client > 0.33.0
gitpython
flask-httpauth
flask-sqlalchemy
flask-admin
flask-bootstrap
flask-wtf
flask-cors
flask
pymysql
tabulate
py-radix
jsonschema
graphviz
ipaddress
deepdiff <= 3.3.0